requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
gevent
gunicorn
langchain==0.0.227
python-dotenv==1.0.0
transformers